+++ Server/Program.cs
@@ -17,7 +17,7 @@ namespace Server
             loop.Start();
         }
 
-        static void Loop(double delta)
+        static void Loop(float delta)
         {
             Multiplayer!.Poll();
             World!.Tick(delta);
